pyproject.toml
[tool.poetry]
name = "module-13"
version = "0.1.0"
description = ""
authors = ["Y3vh3n11 <145870816+Y3vh3n11@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
pillow = "^10.3.0"
fastapi-mail = "^1.4.1"
uvicorn = "^0.29.0"
fastapi = "^0.110.2"
sqlalchemy = "^2.0.29"
libgravatar = "^1.0.4"
python-jose = "^3.3.0"
passlib = "^1.7.4"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
python-multipart = "^0.0.9"
fastapi-jwt-auth = "^0.5.0"
bcrypt = "^4.1.2"
pymongo = "^4.7.1"
redis = "^5.0.4"
python-dotenv = "^1.0.1"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
django-environ = "^0.11.2"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.40.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
